# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/restaurants": {
      "post": {
        "operationId": "RestaurantController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRestaurantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantReturnDto"
                }
              }
            }
          }
        },
        "tags": [
          "Restaurants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "RestaurantController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Default 10",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Default 1",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "foodType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Chinese",
                "Curry",
                "Arabic",
                "African",
                "Turkish",
                "Greek",
                "Kebab",
                "Moroccan",
                "Pizza",
                "Japanese"
              ],
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "operator",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "lesser",
                "equal",
                "greater"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Restaurant"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Restaurants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/restaurants/{id}": {
      "get": {
        "operationId": "RestaurantController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantReturnDto"
                }
              }
            }
          }
        },
        "tags": [
          "Restaurants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "RestaurantController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRestaurantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantReturnDto"
                }
              }
            }
          }
        },
        "tags": [
          "Restaurants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RestaurantController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Restaurants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth": {
      "get": {
        "operationId": "AuthController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AuthController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/self": {
      "get": {
        "operationId": "AuthController_self",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/{id}": {
      "get": {
        "operationId": "AuthController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AuthController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/password": {
      "patch": {
        "operationId": "AuthController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Interview API",
    "description": "RESTful API for use with Interview",
    "version": "2.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:6868",
      "description": "Local server",
      "variables": {
        "port": {
          "default": "6868",
          "description": "The port of the server"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "CreateRestaurantDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 6
          },
          "foodType": {
            "type": "string",
            "enum": [
              "Chinese",
              "Curry",
              "Arabic",
              "African",
              "Turkish",
              "Greek",
              "Kebab",
              "Moroccan",
              "Pizza",
              "Japanese"
            ]
          },
          "address": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "addressLine2": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "outcode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "postcode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "url",
          "name",
          "rating",
          "foodType",
          "address"
        ]
      },
      "RestaurantReturnDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "foodType": {
            "type": "string",
            "enum": [
              "Chinese",
              "Curry",
              "Arabic",
              "African",
              "Turkish",
              "Greek",
              "Kebab",
              "Moroccan",
              "Pizza",
              "Japanese"
            ]
          },
          "address": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "outcode": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "name",
          "rating",
          "foodType",
          "address",
          "addressLine2",
          "outcode",
          "postcode"
        ]
      },
      "Restaurant": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "foodType": {
            "type": "string",
            "enum": [
              "Chinese",
              "Curry",
              "Arabic",
              "African",
              "Turkish",
              "Greek",
              "Kebab",
              "Moroccan",
              "Pizza",
              "Japanese"
            ]
          },
          "address": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "outcode": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "url",
          "name",
          "rating",
          "foodType",
          "address"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "next": {
            "type": "number"
          },
          "prev": {
            "type": "number"
          }
        },
        "required": [
          "totalCount",
          "totalPages",
          "limit",
          "page"
        ]
      },
      "UpdateRestaurantDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 6
          },
          "foodType": {
            "type": "string",
            "enum": [
              "Chinese",
              "Curry",
              "Arabic",
              "African",
              "Turkish",
              "Greek",
              "Kebab",
              "Moroccan",
              "Pizza",
              "Japanese"
            ]
          },
          "address": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "addressLine2": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000
          },
          "outcode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "postcode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "role": {
            "type": "number",
            "enum": [
              1,
              2
            ]
          }
        },
        "required": [
          "username",
          "password",
          "role"
        ]
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "maxLength": 100
          },
          "newPassword": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      }
    }
  }
}